Update the flags needed for Android compatiblity (#1191)
Added a platform check so that the flags will only be set if the OS being run is Android, as they aren't needed for iOS. 

Also changed the original "&&" structure for a simple "if" structure, since the original would incur on a "no-unused-expressions" error on default ESLint configurations. Said error can be overridden by allowing the "allowShortCircuit" option on .eslintrc, but since current guidelines don't recommend it, such cases should be avoided as much as possible, hence the change.